drfcatalog/magazine/serializers.py
from .models import *
from rest_framework import serializers


class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Watch
        fields = ['name', 'description', 'price']


class BasketItemSerializer(serializers.ModelSerializer):
    class Meta:
        model = BasketItem
        fields = '__all__'


class ManufacturerSerializer(serializers.ModelSerializer):
    watches = ProductSerializer(many=True, read_only=True)

    class Meta:
        model = Manufacturer
        fields = ['id', 'name', 'watches']
